Fix #60 Stopped doubling of markers in list
The marker list was becoming size 2*n +1 (where n=size of the marker
list before pulling from the api). This was because we were only pushing
and never clearing. This has been fixed with the addition of a clearing
of markers before we pull from the api. Of course the consequences of
this is that if an api pull fails then we'll have no markers on the map.

We are still experiencing the markers added not toggling though. Working
on this.