Pins and Comments relationship update
Updated Pins to create a new comment when they're being created. This
needed the comment types and the pins type to match up better, so the
pins type have replaced some of the comments type. The documentation
will need to be updated to reflect this.

The id that is returned with
comments is from Pins.key().id(), I don't know if we ever use this
anywhere, however if we implemented being able to jump to a pin from the
comments page, then this id would probably be a place to start.