Fix #46 Moving api variables to be local
Moving api endpoint variables from their more global scope of the
APIConnector.prototype into the functions they are used in. I am curious
if this is neccesary because the variables are strings and I'd guess
that since all the strings used in the expressions they are used in are
constants, that the javascript engine could optimize them a bit.

Also, the pushHeatMap data function had the ajax code wrapped inside of
its if(window.logging) statement. This seems wrong so I have corrected
it. However to make sure, I'd like @JoshuaDickerson to take a quick
look.