Fix return type problem with fragments
The fragment sections were of type android.app.Fragment instead of the
type specified in MainActivity (android.support.v4.app.Fragment). I'm
not sure if any other imports will have to change but if you get a
strange return type error where you know a subclass should be able to be
returned as a member of it's parent class and it's not working. Chances
are it's an import error.

Also, case 2 is now returning a FeedSectionFragment since this was my
test case. According to the documention on Fragment, any subclass should
include an empty constructor with no arguments, so this has been added
to the FeedSectionFragment as well.