Temp Fix #64 and offset integer relaxation
For the time being, the REST API heatmap endpoint supports the POST verb
and performs the same operations as a PUT request. This is to handle
inadequacies in the FSNetworking library which does not support PUT
requests and simly states to do it yourself in the readme. This is a
temporary condition for the API until the FSNetworking library has been
extended.

Offsets no long need to be integers and can be decimals as well. This is
done because the mapkit in apple gives back offsets in sub-degrees which
is decimals. So we need to support decimals as well as integers.