Fix #82 Integer for empty pin in comments
As Evan pointed out, the empty string being returned for a pin when
retrieving comments (due to comment not having a pin associated with it)
is inconsistent with the API specification. Therefore a pin of 0 will be
returned if there is no pin associated.

A 0 is an alright flag value since the pins come from a reference
property which is an key id generated by the datastore from the contents
of the pin itself. Therefore we will never actually have a pin with an
id of 0.

If the UI relies on the pin property to jump from a comment to its
associated pin (not implemented yet, but is an enhancement idea) then
this value should be checked to be non 0 before attempting to follow a
pin to avoid errors.