Fix #72 Non-empty pin message check
The pins POST endpoint now requires the message being sent to be
non-empty. If the pin's message is empty a Semantic Error will be raised
and a 422 returned to the client with a short error message indicating
their message was unprocessable due to the message being blank.

Update to readme to reflect change as well as addition of details to
the two types of error codes returned by the API.