Updated resource path. Memcache stratey update
References: #84 #88

The resource path within the constants file was missing a forward slash,
so I have added this. Also, removed the write key and the strategy for
updating the memcache after every 20 writes to the database. Instead, we
now invalidate the cache on every write to the comments endpoint.

This is done because the frequency of reads will be higher than writes, so
the cost of creating pages of comments only to have them be invalidated
by the cache is less of an impact than one might at first think.

Note that only the comments endpoint supports the memcache right now.