Raw support for heatmap and documentation
References #88

Heatmap GET request now supports a raw parameter. If given this as true
then the data returned will have its secondsWorked field be the actual
seconds worked and not a scaled value. This is for information purposes
primarily and could be used to generate reports about the number of
people working and that like.

Added documentation to API specification detailing the use of the raw
argument that can now be passed to the heatmap endpoint.