Added list of errors returned by API
References: #88

Added list of error messages to the specification so that clients can
handle all errors gracefully and be able to check the message of the
various errors in order to display whatever dialog neccesary to inform
the user that there was a problem.

A reminder to myself for later, when performing the next release of the
API I should update the version number of the API within both the yaml
file and the returned JSON from the get request api.py.