Updated API to use ancestor queries
References #88

Updated Datastore to use all ancestor queries instead of regular
filters. This ensures that the datastore will be strongly consistent, as
queries to the same entity group are the only ones upon which strong concistency
as opposed to eventual consistency) is enforced. We limit the results of a query
to the specified entity(defined by the app_key method) and its descendants.

Signed-off-by: Phelan Vendeville <phelan.vendeville@gmail.com>