Added background to actionBar
Added background to the action bar. It must be set via the stacked
background drawable function otherwise, it will not override the
backgrounds of the tabs themselves.

Also, we need to figure out how to tile the background of I will have to
use imagemagick/rsvg-convert during the image creation script to
generate appropriately sized images for the backgrounds for tiled
objects.