Tiled background for actionbar
There is now a repeated image for the action bar. Unfortunately I had to
mess with the bottom_menu.svg file used to generate the image because it
has extra space which it really shouldn't. I will have to talk to Doug
about this as I should not have had to spend over an hour in inkscape
messing with this.