Fixed up bottom image for comments
Work on #93

Set the bottom image to fill_parent instead of wrap content. Not sure
why it was set otherwise. Also determined the reason the comments look
so huge and gross. It's because the svg images are too high and have
tons of white space. This is very unneccesary and causes badness all
around for us. I will have to email doug and request new SVG's.