Fixed up SVG's in inkscape
Work on #93

Bottom bubble svgs will need to be resized to 342.60 by 22.24 from
document properties.
Top bubble svgs need to be resized to 343 by 10.

Regenerated the images from the svg's as well. Not sure what to do for
the center bubble yet as it really just looks like a blank image.