Added reversal of comment bubbles and margins
Work on #93

Added margins around the comment boxes as well as alternating the
bubbles direction on each one so we get that conversation feel to it and
the feed doesn't look bland.

Regenerated images from svg's for the bubbles. Still need to work on
getting the center piece to tile correctly.