Added twitter-esk timestamps to comments
Work on #93

Added date time parsing and appending the timestamp to the textview
within each comment on the feed page. It's probably not the most
efficient code, since it has to do anything with Java's terrible and
unfriendly time libraries. But it will do for now.

The feed page loads very slow. I believe it might be best if I attempt
to figure out how to do some of the loading on the page asynchronously.