Faster loading time for Feed
By delegating the list adapter to perform it's job of linking the
comments from the API to an Asynchonous task, the load time of the
comments page is much faster. Before this was done async, the load time
to generate the view for the page was on the order of 3-5 seconds
depending on the connection of the internet. The API call performs it's
task asynchronously and therefore is not the problem. Thus leading me to
conclude that the list adapter itself, and its corresponding view
rendering was the slow problem. By imperative results of the feed page
now loading within 1 second of clicking, I can semi-confidently say that
the problem was indeed where I suspected it was.