Fix NPE for Timestamp parsing
If a Comment is made programatically, or does not have a timestamp set.
The timestamp parsing will fail miserably due to a null pointer.

To correct this we have to check the timestamp whenever we try to grab
it from the comment's .getTimeStamp() function. We check it in code, but
it may also be a good idea to verify the values of the various keys when
creating the json in the API connector.

@primehunter326 what do you think about validating the input from during
your JSON parsing to avoid null pointers? Should we put in default
values or should we allow the NPE's to surface and deal with them during
runtime? It makes sense to me to place default values since it will
avoid system crashes, however we might also get funky behavior depending
on the default value. A good default value for the timestamp if one is
not supplied could be the current time. Which is possible through just
the new Date(). To put it into the format the view rendering will
expect, you can use:
 new SimpleDateFormat("E LLL d kk:mm:ss yyyy",Locale.US);
And I'm expecting it to be GMT time when I'm rendering it (I convert it
to EST)