Got a basic text input on.
Working on #93

Got a EditText appearing after fiddling with the size of the list view.
Which was taking up the whole page. I don't know if the current values
being set through the graphical designer in Eclipse are good. There
might be a better way to set the properties so they're not fixed at some
dp size.