Initial Spec on debug endpoint
References #100

Updated specification for api to include a possible scheme for the debug
endpoint's interface. This is subject to change upon discussion with
@the-hobbes

The error codes portion of the specification is not done and probably
won't be done until we decide on the spec and the validation code is
written. Currently, the endpoint functions for GET and POST requests
exist but does nothing. The DELETE handling functions are not
implemented yet. POST has some small degree of validation but nothing
endpoint specific yet (only json validation).

We need to determine our strategy for dealing with authentication and
deletion of messages. We'd like to allow a client to delete their own
messages if they don't want to be transporting that data to us (or be
able to shut off reporting on their end entirely if we feel its a legal
issue. ) Currently the spec states an origin/master key policy. Where we
can use salt + device id to determine the originator of the debug
message and use that as part of validation before deleting a message.
Also, a master key for developers could exist. Either that or we just
allow devs to delete reports through the app engine console. But this is
horribly slow and it'd be nicer to automate it.