Example delete request for debug endpoint
Reference #100

Added example of sending a DELETE request. Note that the DELETE verb
does not take its arguments in a message body, but in the url itself.
This is according to the HTTP spec. Since we are deleting simply by a
resouce identifier, aka, the URI. It might be more semantically correct
if we had the url be /api/debug/origin?hash= but I'm not sure how to do
that with python and placing both into the DELETE url parameters should
work just fine.