Work on Enhancement #101 (status codes) on heatmap
Implementing the changes discussed on github issue #101. Including a
status_code key into the response from the API so that all information
is within the json as well as the HTTP headers.

Implemented changes on heatmap endpoint and updated spider testing to
assert that this is true. Spider passes, so it's working.