Work on Enhancement #101 (status codes) heatmap.
Continued work on heatmap endpoint. This is a breaking change for the
API clients because it modifies the heatmap to return a json object
instead of a json array.

The new json object has keys of status_code and grid. The value of the
grid key is the old json array that the clients using the outdated API
will be expecting.

An idea has just occured to me. I will update just the appspot api
site and clients can use that to test their work against the new API
while still using the old version from the app site.