Work on enhancement #101 (status_codes) pins
Implementing the changes discussed on github issue #101. Including a
status_code key into the response from the API so that all information
is within the json as well as the HTTP headers

This is a breaking change to the pins endpoint. The json response
returned will now be an object, not an array. The object has both a
status_code and a pins key. The pins key value is the array that was
returned before this change occured. The status_code is the HTTP
response code also included in the headers.