Work on #110
@primehunter326 @milus16

It seems like the issue has something to do with the adapter getting
ready to load the next page and trying to optimize memory or something.
Looking at the stack overflow I put onto the issue suggest to use a
TabAdapter.

I've implemented one here. But it's being troublesome. It does load up
the right pages on the right tabs being clicked no problem. So it would
resolve the issue, however the view is not being displayed. If we can
get the view to display on this branch then we could modify the main
android branch to use it and close the issue.

Anyone more familiar with the tab adapter then I am? I can't really see
why the view wouldn't be instantiated here. The documentation where I
got the sample code from is here:
http://developer.android.com/reference/android/support/v4/view/ViewPager.html