Extensions to spider for delete endpoint testing
References #100

Working on extending the spider for the debug endpoint. Added testing
and validation for POST and GET endpoints of debug endpoint. All is
working as expected. I foresee some trouble with spidering the DELETE
request only because we need the datastore to be operational in order to
properly validate it.