Updated spec on debug endpoint
References #101, #100

Updated spider to perform all but one check (the check is not possibly
until the abstraction layer is complete).

Updated spec with note about status code 204 and the DELETE request
response being an empty string. How to handle this is listed as a
comment by myself on issue #101 on github.