Updated datastore to handle debug
References #100

Added origin field to debugMessages. Added methods to debugReports by
their origin, hash, and since time.

The biggest change is to the paging function. Instead of being strictly
limited to using the comments. It now takes a few extra parameters and
handles the paging properties for both the debug endpoint and the
comment endpoint. While coding this, I was thinking perhaps we should
extract the paging strategy into its own class or maybe a function that
takes a list of items to create the paging for and then returns it up to
another function that handles the formatting and endpoint specific
features.

Posting to the debug endpoint now creates a debugReport object as well.
So that portion is functional. Only responding to the DELETE verb needs
to be done.