Work on DELETE for debug endpoint
References #100

Hooked the debug controller and the datastore together through the
abstraction layer. The datastore is not deleting anything yet. But once
it does the test in the spider should be uncommented since it's
functionality testing for 404 and 204 will be being returned.

The spider is not currently passing for some reason. Something to do
with the since time. I will look into this next while @the-hobbes
finishes the DELETE request functionality.