Updated Spider for 404 handling and debug endpoint
References #100

Updated spider to handle a 404 with its better response handler class.
The spider will not successfully execute because items are not deleted
from the datastore yet.

Updated initial paging to handle sinceTime like paging does. And
updating paging to add sinceTime into the key's for the memcache in
order to properly handle the debug Reports