Refactoring Status Codes
References #101, #100

Preparing for new release of API. Enhancements #100 and #101 are slated
to be released tomorrow by my internal timeline.

All spider tests pass right now, but we should refactor everything to
use the standard ERROR_STR constant and string interpolation for
returning error messages for endpoints. I've done the comments one in
this commit. It would be nice if @the-hobbes could take on pins or the
heatmap.