Added yellow bubble bottoms and removed old svgs
Removed the old folder for svg's if we need one of them you can always
check out the commit before this one (the merge) and copy the file over.

Added the yellow bubble bottom svg's by copying/modifying the red
bottoms and changing the fill color of the markup. Gave me the idea that
we could probably condense the repository size a bit more by having
bubble_ templates and then doing a string replace for the color. This
would shorten up the create_images script a bit. Taking 12 lines down to
3. Considering this for refactoring later on.