Updating svg's (part 2)
Grabbed bottom_menu svg from old commit. Green background for the menu
on android is now showing up (mostly) nicely. It's not as good as what's
on the android branch but we'll have to play around a bit more and try
to figure out exactly what changed with the file.

The key thing is that the bottom menu svg that was changed was not wide
enough would be my guess. We need at least 1-2 pixels wide for the image
otherwise it seems that it won't render properly when given as a
drawable to the device. That results in a weird white background image
instead of the green one we expect.