Fixed svg "blinds" effect
The svg's had extra whitespace in them that caused the comments to have
horizontal lines accross where they were tiled leading to a sort of
window blinds affect.

To fix this, the svg's were simply stretched so they covered all of the
area of the canvas they exist on. Then the images were regenerated.