Added more logging to find #110
Added more logging into the methods being used that might be causing
issue #110

http://stackoverflow.com/questions/13214063/getitem-calling-multiple-time-in-fragmentpageradapter?rq=1

That stack overflow seems to be what we're experiencing. Multiple calls
to get item when clicking a tab. It seems like the android os is
pre-rendering the next frame, which unfortunately, causes our network
connects to boot up and fire away. This is unfortunate to say the least.
We could instantiate all the fragments and keep references to them on
application start up and make sure not to load any data into them. And
then when we detect the view is the fragment we could ask it to load
everything. This seems like it would work, but also knowing android, it
seems like we're just as likely to get null pointers and inflation
problems with views not being present when we're trying to add
information to them. A major refactor is needed of the android code. The
API code needs to be completely asynchronous. Each Fragment needs to be
stubbed out so it's view can be made without reliance on the data being
placed into it, and then methods to populate the data itself need to be
created and used. You'd think we could use the onCreateView, but I'm
pretty sure the pre-rendering is hitting that too. We're probably going
to have to call methods on each fragment to populate from the page
handler selecting the item itself.

@primehunter326 What do you think?

Also I commented out the logging code for the comments, I was tired of
scrolling past every comment's information.