Removed random folder.
Don't know where this folder came from, I could go a git blame and
bother whoever's fault it is that these random folders came into
existence; but I'll hold off.

@popwarfour I'm just going some cleaning of the repository and there's a
bunch of folders in the ios directory and there's one GoGreen folder in
the root level of the repository. WHICH of these is the build you're
using, and please use git rm to remove the other ones and have the ios
project be just in the ios directory. Thanks.