Added secondsWorked long and map.xml
Added secondsWorked variable to MainActivity. Why in MainActivity?
Because the main doesn't go out of scope so I'm not going to lose my
variable when switching tabs. If I had internet right now I would be
looking at persistence instead and how to use the android local storage
to keep track of the secondsWorked. We're going to need something like
the following:

secondsWorked   (actual time user has had timer on app)
secondsSent	(time that has been sent to the server so far)
lastTimeSent	(time since last sending)
lock		(boolean mutex of some kind)

And it's simple from there. If the secondsWorked > secondsSent and we're
in one of the points in the app's lifetime where we send data, we'll
send the secondsWorked up. If the lastTimeSent is some arbitrary unit
away (like 5 minutes) then we'll initiate a connection to the API's
heatmap endpoint and send our data up. Upon a successful status code
we'll have secondsSent incremented by how many seconds we sent up.

Now here comes the key point. Because all our stuff is asynchronous in
order to not slow down the UI, we need to only be sending data to the
heatmap once every arbitrary unit. So we need to have some classwide
boolean lock to prevent race conditions. Otherwise we're going to be
confused when we try to send our data because the adapter is on during a
comment submission and we also happen to hit that arbitrary unit of
time.  Both requests will try to write to the class variables and
ka-poof. There goes out chronological integrity.

Adding map.xml because it's been sitting in my repository for forever
and I'm tired of looking at it in the untracked list.