Pulled in old source code example for storage.
Pulling in an example storage file from a project I did for Dr. Snapp.
Simple sqlite helper for storing tasks. This will be a good thing to
modify for use with the heatmap data / secondsWorked.

We're going to need some type of hook into the android's lifetime so
that we can monitor and send data off. I need to look into the android's
lifetime in the documentation to find out what to do here. Sadly that's
difficult to do on the bus when the wifi has cut out.