Updated timer to stay semi persistent
The timer will now state correct during runtime of the application. The
storage has not been implemented yet (sqlite) but once it's done then we
can use that to store the time on application exit/pause. We can also
then use the stored time during the onCreate method of the
HomeSectionFragment to set the HomeSectionFragment.currentTime variable
to be whatever the correct number of seconds are.

The table schema for the timer's sqlite tracking still needs to be done.
As well as gps polling data that we will send to the server. Perhaps we
could simply store a serialized heatmap gridpoint for each unique point
on the gps we're located at (with some precision of rounding to lump
points together for storage space)

@primehunter326 Your thoughts on the heatmap storage structure? We still
need a map and to poll the gps itself, but it's not terribly hard to do
the polling if I remember correctly, it's just the map that seems
difficult to me. To divy up the android tasks a bit, do you want me to
work on getting the persistent stuff and the gps things going and you
continue working on the android-menu stuff as well as possibly looking
into how to stick a map into a fragment? I don't care that much about
getting heatmap markers onto the map or having any functionality
besides: Hey here's a map that won't crash the app if we don't have
network connectivity. I just want the application to be reaching a more
'finished' looking state at least even if some of the functionality
isn't there (I'm sick of the test button page and want to delete the
dummy fragment code.)