Began Adding persistence to home section.
It's perstent on exiting the application then coming back to it. But not
for switching to the third tab and then back to the first. What we need
to do, I think, is also enable saving the same when switching the tab
and the HomeSectionFragment's fragment gets killed for memory (I know it
is because it has to reload it when we switch back to it)

It was persitent the other way around before becuase of the classwide
variable, so either I check to see if the classwide variable is != 0 and
then pull from the database, or we add in saving when moving between
fragments / when the fragment activity is paused.