Change comment/pin types to use just 3 types.
The comment and pin types are no going to just be MARKER, COMMENT or
ADMIN. And that's it. This commit updates the constants as well as the
readme for documentation purposes (RTFM&&STFU||GTFO) and also fixes
the spider to run correctly given the new circumstances.

This commit is part of preparation for a release of API version 3. Which
will have these new types and will also include:

- Types of just MARKER, COMMENT and ADMIN
- Proper entity counting as per @the-hobbes
- Deletion of comments support through the api by comment id
- Deletion of pins support through the api by pin id (and deletion of
  any comments associated with this pin)

For development purposes the appspot domain at greenupapi.appspot.com
will switch to using version 3's in progress version. You can see what
version of the api you're using by querying the /api endpoint and noting
the version number. The stable version is 2-5 so the
greenupapp.appspot.com will continue to serve that version until the
above features are fully implemented.

If you are pulling down this API version to your local machine you will
_need_ to clear out your old database otherwise you will experience
server errors when asking for comments/pins. If you ask me for
assistance I will point you to this commit and kindly tell you to read
this nice commit message I took the time to write for you.