Work on deleting Pin by ID. Reference Error.
Having trouble with the reeference property and deleting that. It
doesnt' seem like filtering to get comments based on the pin is working,
I also tried using the pinID but to no avail.

@the-hobbes could you take a look at this? Is it possible to do
something like pin.comments.delete() or something of that nature? For
example, rails has Association.clear / destroy_all which clears our
associations from an object itself. I imagine the app engine must have
something like that that we could use? Or some type of dependent destroy
that automaically is fired off from dekeintg the pin or not to.

To reproduce the error:

- run git fetch origin referenceProperty
- run get checkout FETCH_HEAD
- run the spider, it will pass.
- Run it again. It will fail on retrieving the comments.
- fix the problem if you can ?
- if you fix it: git checkout -b referenceProperty
- git push origin referenceProperty
- send me a pull request or push direcly to the mainline's
  referenceProperty branch or merge it with --no-ff to
  feature/type-clarification branch (which is essentailly the api
  version 3 branch)