Fixed the reference Property issue
I am so infuriated with the GAE. This is an ugly hack to fix the broken
reference property. Trying to retrieve the comment that will be broken
when deleting the pin doesn't work. I tried:

filter(pin = the id), filter (pins = the id) doing a ,get() and deleting
just that, not doing a get and iterating and deleting, doing a get and
iterating, etc etc. Nothing seems to work to delete the referenced
comment. So fuck it. There's a try except in the comment creation now
that fixes the issue. It's an ugly stupid hack.