Fix #129
The problem was that we're all morons. I've corrected this with a heavy
dose of sarcasm, hatred, and far too many cups of Tea. Enjoy.

Seriously though, there is no id! there is key() and there is the id()
of the key object. If we were in RoR then there would be an id, but
sadly that would make too much sense. Although, if we were in RoR then
we wouldn't be doing a raw json API server anyway, because that's not
what RoR is designed for and it doesn't play well with raw HTTP. Not too
mention it's got a bloated rendering and asset pipeline.

Anyway, my dayjob aside. This is the fix. We need to stop pretending
that the GAE makes any sense with it's ORM.