OOP approach on chronotime from database
Added a new ChronoTime object that is WAY better than returning an array
of strings. I think I must have been tired when I wrote that other code.

While as of this commit the time computation is WRONG and not quite
right (14400 seconds? what?) the important part is that it doesn't
crash the program anymore when shutting down the app or switching
between tabs. OR at least not that I've seen.