Comment filter toggles now take effect as soon as the drawer is closed
It's probably not as efficient as it ought to be because it basically 
rcreates the CommentAdapter each time, but this way is relatively simple
and could be changed.

The fragment now keeps a reference to the comment page object returned 
by APIServerInterface.  I was originally going to just have it keep a 
reference to the ArrayList of Comments but determined this wasn't ideal 
for a number of reasons.  Because the comment page is returned by the 
CommentLoadTask when it is done executing in its separate thread, the 
drawer is locked closed initially when the fragment displays and then
unlocked when the task is finished.  This is to prevent a scenario where 
the user might trigger a refresh of the filtered view by closing the
drawer, resulting in a NPE.