Commiting what I had -- Nonworking code
The chronometer isn't working right yet. I'm not sure if I can get it to
without holding some state in the chronometer itself. I may very well
implement a new class to handle the chronometer because honestly, it
seems like a much better idea than hammering away in the
HomeSectionFragment.