Work against #118 "2nd set of Eyes"
References #118 and #130

@the-hobbes
The monkey patching is cool, but is it neccesary? I've taken it out for
the same of clarity. (I'm sorry if you spent a lot of time on it, it is
very cool how it allowed us to not have to explicitly call anything --
really)

I'm calling the increment manually and it's giving me the right count
for all entities except the gridpoints. For some reason, the gridpoints
show up multiple times in the database and I'm not sure why. If you're
ok with us un-patching the datastore you can merge this commit in with:

    git fetch origin entityCounting
    git merge --no-ff FETCH_HEAD #I'm assuming you're on develop

And it should take. I think there might be more code I could take out,
as the 'after put' and 'before put' logging no longer appears now that
I've removed the patching.

If the monkey patching worked properly it'd definitely be the way to go
but the race condition and asynchronous code is definitely the problem I
think.