Working Chronometer besides some formating issues
Turns out the lack of persitence was because (surprise) the database was
being deleted each time the view was created, this resulted in a less
than desirable behavior.

The only strangeness now is that the format of the string doesn't appear
correct until the tick event fires. This means that if the chronometer
is off when switching back to the tab, the display reads 00:00:00 until
you click start on which it will load the proper amount from the
database.