Chronometer state correct within application
Chronometer when resuming application is wrong and has not been
counting. This is because the stopped time is not being taken into
affect. However, the chronometer is working correctly (besides one
slight bug) in the application itself.

The _slight_ bug is that an extra 0 appears in the seconds area on
occasions when resuming the view from another panel. Why this occurs
probably has something to with the text being updated while the format
of the text is being computed. It doesn't happen often though and fixes
itself after 1 second of ticking to it's not really that big of a deal.