Added proper pause time calculation
When loading the application from the screen, if the chronometer was
active when the application was shutdown the time between shutdown and
restarting the application is added to the secondsWorked.

Something that happens infrequently is it taking two clicks to get the
start/stop button working again after resuming the application. Not sure
if this is the emulator just lagging and not noticing mouse clicks or if
it's something else.