Implementation of Feature #132
References #132 and #127

@popwarfour

The addressed field is now returned within the comments. It is returned
as false if there is no pin associated with the comment. So it is the
clients responsibility to check if the pin is "0" or not before making
any assumptions on the addressed field.

This is implemented on the version 3 branch and should NOT be merged
into the develop branch until @the-hobbes is done with the entity
counting bugs, after that (and any other feature requests that are
plausible) then the API version 3 will be released.

As I noted in #127 greenupapi.appspot.com _will_ server the in progress
version of the API if you want to be testing against it.