Fixed response from pins filtering
Pins filtering was returning the case I never thought it would hit. This
is because of Python's weak typing. It intepreted 0 latitude or
longiture as false and managed to hit that case. Returning a string to
the ios device that thought it was in Africa.

This has been fixed. And I blame Africa.