Re-arranged the layouts
This is so all of the edit controls are in one layout which can have its
visibility toggled.  

Sidenote: when you rearrange layouts, clean the project because eclipse
apparently doesn't make the necessary changes to R.java automatically.
When I first tested this the app crashed with a
IllegalClassCastException because it thought I was casting a
relativelayout to a spinner.