Feature Request #138
This is a simple implementation of the feature requested by @popwarfour

There is hardly any error checking involved and submitting a non
integral value will break and cause a 500 server. This error checking
code will come next. However I am pushing this now so that more work can
be done. Note that this is only implemented on the type-clarification
branch.