Error Checking for Feature Request #138
Integral values will not be accepted by the endpoint. If attempted a 400
error will be returned with a message of "Pin Id must be an integral
value"

The other error message that can occur is a 404 if the requested pin has
not been found. In which case an error message stating as such is given
to the calling party.