Initial implementation of heatmaps
Added the android maps utility library which lets you display
heatmaps as a TileOverlay on the map given a set of LatLng
objects.  It is currently set up to use some test data that
is stored locally as a 'raw' resource.  The data is straight
JSON from the API so the only extra thing that had to be added
was a function to get all of the LatLng objects as an array.

I had to comment out a compile directive from build.gradle because
the support library was getting included twice.