Heatmap data is now pulled from the server instead of the test data
The API interface is set up to use the old appspot domain so there
is hardly any data (I thought it was broken at first).  Tried
changing the URL to the correct domain but that threw SSL exceptions
so it will need to be fixed.  It is also set up to use a separate
asynctask as the one it uses for pins, even though they're basically
the same.  They could be refactored into one task that takes a parameter
that tells it what to fetch.